+++ app/_layout.tsx
@@ -1,12 +1,13 @@
-import FontAwesome from '@expo/vector-icons/FontAwesome';
 import { DarkTheme, DefaultTheme, ThemeProvider } from '@react-navigation/native';
 import { useFonts } from 'expo-font';
 import { Stack } from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
 import { useEffect } from 'react';
 import 'react-native-reanimated';
+import FontAwesome from '@expo/vector-icons/FontAwesome';
 
 import { useColorScheme } from '@/components/useColorScheme';
+import { FontProvider } from '@/components/FontContext';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { PersistQueryClientProvider } from '@tanstack/react-query-persist-client';
 import { createAsyncStoragePersister } from '@tanstack/query-async-storage-persister';
@@ -42,7 +43,20 @@ const asyncStoragePersister = createAsyncStoragePersister({
 export default function RootLayout() {
   const [loaded, error] = useFonts({
     SpaceMono: require('../assets/fonts/SpaceMono-Regular.ttf'),
+    // Amiri fonts
     'Amiri-Regular': require('../assets/fonts/Amiri/Amiri-Regular.ttf'),
+    // Playpen Sans Arabic fonts
+    'PlaypenSansArabic-Regular': require('../assets/fonts/Playpen-Sans-Arabic/PlaypenSansArabic-Regular.ttf'),
+    // Noto Sans Arabic fonts
+    'NotoSansArabic-Regular': require('../assets/fonts/Noto-Sans-Arabic/NotoSansArabic-Regular.ttf'),
+    // IBM Plex Sans Arabic fonts
+    'IBMPlexSansArabic-Regular': require('../assets/fonts/IBM-Arabic/IBMPlexSansArabic-Regular.ttf'),
+    // Noto Kufi Arabic fonts
+    'NotoKufiArabic-Regular': require('../assets/fonts/Kufi-Arabic/NotoKufiArabic-Regular.ttf'),
+    // Baloo Bhaijaan 2 fonts
+    'BalooBhaijaan2-Regular': require('../assets/fonts/Baloo-Bhaijaan/BalooBhaijaan2-Regular.ttf'),
+    // Noto Naskh Arabic fonts
+    'NotoNaskhArabic-Regular': require('../assets/fonts/Noto-Naskh-Arabic/NotoNaskhArabic-Regular.ttf'),
     ...FontAwesome.font,
   });
 
@@ -66,7 +80,9 @@ export default function RootLayout() {
       client={queryClient}
       persistOptions={{ persister: asyncStoragePersister }}
     >
-      <RootLayoutNav />
+      <FontProvider>
+        <RootLayoutNav />
+      </FontProvider>
     </PersistQueryClientProvider>
   );
 }